+++ estimate_info.php
@@ -27,15 +27,22 @@
                 if (isset($_GET['price_edited_estm'])) {
                     $_SESSION['estm_arr'][$_GET['index_edited_estm']][2] = $_GET['price_edited_estm'];
                 }
-                if ($_POST['description'] !== null) {
-                    $_SESSION['estm_arr'][$_SESSION['i_estm']][0] = $_POST['description'];
-                }
-                if ($_POST['price'] !== null) {
-                    $_SESSION['estm_arr'][$_SESSION['i_estm']][1] = $_POST['qty'];
-                }
-                if ($_POST['qty'] !== null) {
-                    $_SESSION['estm_arr'][$_SESSION['i_estm']][2] = $_POST['price'];
-                }
+				
+				if (isset($_POST['description'])) {
+					if ($_POST['description'] !== null) {
+						$_SESSION['estm_arr'][$_SESSION['i_estm']][0] = $_POST['description'];
+					}
+				}
+                if (isset($_POST['price'])) {
+					if ($_POST['price'] !== null) {
+						$_SESSION['estm_arr'][$_SESSION['i_estm']][1] = $_POST['qty'];
+					}
+				}
+                if (isset($_POST['qty'])) {
+					if ($_POST['qty'] !== null) {
+						$_SESSION['estm_arr'][$_SESSION['i_estm']][2] = $_POST['price'];
+					}
+				}
 
                 if (isset($_POST['submit'])) {
                     $_SESSION['i_estm']++;
@@ -68,24 +75,26 @@
                                     <td colspan="2" align="center"><input type="submit" name="submit" value="Add"></td>
                                 </tr>
                 ';
-                for ($i = 0; $i < sizeof($_SESSION['estm_arr']); $i++) {
-                    if ($_SESSION['estm_arr'][$i][0] !== null) {
-                        echo '<tr bgcolor="#c4daff"><td>'.$_SESSION['estm_arr'][$i][0].'</td>';
-                    }
-                    if ($_SESSION['estm_arr'][$i][1] !== null) {
-                        echo '<td>'.$_SESSION['estm_arr'][$i][1].'</td>';
-                    }
+				if (isset($_SESSION['estm_arr'])) {
+					for ($i = 0; $i < sizeof($_SESSION['estm_arr']); $i++) {
+						if ($_SESSION['estm_arr'][$i][0] !== null) {
+							echo '<tr bgcolor="#c4daff"><td>'.$_SESSION['estm_arr'][$i][0].'</td>';
+						}
+						if ($_SESSION['estm_arr'][$i][1] !== null) {
+							echo '<td>'.$_SESSION['estm_arr'][$i][1].'</td>';
+						}
 
-                    if ($_SESSION['estm_arr'][$i][2] !== null) {
-                        echo '<td>'.$_SESSION['estm_arr'][$i][2].'</td>';
-                    }
-                    if ($_SESSION['estm_arr'][$i][0] !== null) {
-                        echo '<td align="center"><button><a href="edit_estimate_detail.php?description='.$_SESSION['estm_arr'][$i][0].' &qty='.$_SESSION['estm_arr'][$i][1].' &price='.$_SESSION['estm_arr'][$i][2].' &index='.$i.'">Edit</a></button></td>';
-                    }
-                    if ($_SESSION['estm_arr'][$i][0] !== null) {
-                        echo '<td align="center"><button><a href="edit_estimate_detail.php?index_deleted='.$i.'">Delete</a></button></td></tr>';
-                    }
-                }
+						if ($_SESSION['estm_arr'][$i][2] !== null) {
+							echo '<td>'.$_SESSION['estm_arr'][$i][2].'</td>';
+						}
+						if ($_SESSION['estm_arr'][$i][0] !== null) {
+							echo '<td align="center"><button><a href="edit_estimate_detail.php?description='.$_SESSION['estm_arr'][$i][0].' &qty='.$_SESSION['estm_arr'][$i][1].' &price='.$_SESSION['estm_arr'][$i][2].' &index='.$i.'">Edit</a></button></td>';
+						}
+						if ($_SESSION['estm_arr'][$i][0] !== null) {
+							echo '<td align="center"><button><a href="edit_estimate_detail.php?index_deleted='.$i.'">Delete</a></button></td></tr>';
+						}
+					}
+				}
 
                 echo '
                             </tbody>
